fix(nc-dav): preview-generation queue binds INT4 columns as i32 on Postgres
oc_preview_generation.file_id/queued_at are `integer`; the strict native
driver (T3.3, 576362f) rejects i64 binds there, so every file write's
queue INSERT failed at warn-level and background preview generation was
starved — new files never received previews.  SQLite keeps the i64 bind
(INTEGER is 64-bit).  Verified: a PUT through the SUT now lands a row in
oc_preview_generation.

+++ core-rs/crates/nc-dav/src/preview_queue.rs
@@ -35,11 +35,14 @@ pub(crate) async fn queue_preview_generation(
          )",
         prefix = prefix
     );
+    // The columns are `integer` (INT4) on Postgres — bind at their exact SQL
+    // types or the strict native driver rejects the statement (every file
+    // write then fails to queue, starving background preview generation).
     let result = match pool {
         DbPool::Pg(p) => sqlx::query::<sqlx::Postgres>(&sql)
             .bind(uid)
-            .bind(file_id)
-            .bind(queued_at)
+            .bind(file_id as i32)
+            .bind(queued_at as i32)
             .execute(p)
             .await
             .map(|_| ()),
